### 2017-07-21 #322 @whedon
[PRE REVIEW]: BayesianNetwork: Interactive Bayesian Network Modeling and Analysis
**Submitting author:** @paulgovan (<a href="http://orcid.org/0000-0002-1821-8492">Paul Govan</a>)
**Repository:** <a href="https://github.com/paulgovan/BayesianNetwork" target ="_blank">https://github.com/paulgovan/BayesianNetwork</a>
**Version:** v0.1.3
