### 2017-07-22 #322 @katyhuff
@janschulz @cdcrabtree @wrathematics -- I believe you have the background to review this submission -- are any of you interested in reviewing this submission for JOSS?
