### 2017-07-23 #322 @janschulz
I've no clue about bayesian models, so if that's not a problem I can do the review.
