### 2017-07-23 #322 @katyhuff
I see. Thanks letting me know. I'd like for someone with bayesian
experience to handle this review if possible. @cdcrabtree or @wrathematics,
do either of you have experience with bayesian models?
