### 2017-07-31 #322 @katyhuff
@cdcrabtree : Would you be willing to review this submission for JOSS?
