### 2017-08-01 #322 @katyhuff
@wrathematics : Would you be willing to review this submission for JOSS?
