### 2017-08-03 #322 @cdcrabtree
@katyhuff Sorry about the late reply. After looking over the package, I don't think that I have the background knowledge necessary to be a good reviewer here.
