### 2017-09-25 #322 @katyhuff
@paulgovan , Please don't think I've given up. I'm having a terrible time getting folks to agree to review this. Many have cited that it is beyond their comfort zone in machine learning. Thank you for your patience. I am continuing to contact folks.
