### 2017-10-19 paulgovan/BayesianNetwork#6 @rgiordan
Unit tests
This issue is for https://github.com/openjournals/joss-reviews/issues/425

There doesn't seem to be an automated way to test whether the shiny app is working as expected.  It seems to me that it would be good to have some tests to check, for example, for breaking changes to the shiny or bnlearn API.

I appreciate that it's not always easy to test web interfaces, and I'm not sure what would be best -- I haven't developed in shiny myself!  But I did find [RSelenium](https://cran.r-project.org/web/packages/RSelenium/vignettes/shinytesting.html), which says that solutions are out there.
