### 2017-10-19 #425 @rgiordan
The package is missing automated tests.

https://github.com/paulgovan/BayesianNetwork/issues/6
