### 2017-10-19 #425 @rgiordan
@katyhuff

Sorry, this is the first time I've done this.  Where do I find the software paper?  I don't see it in the repo, nor in any of the issues or emails.
