### 2017-10-19 #425 @katyhuff
@rgiordan Thanks for the question. This varies by repository. It looks like the paper for this one is at : https://github.com/paulgovan/BayesianNetwork/blob/master/inst/paper
