### 2017-10-20 paulgovan/BayesianNetwork#7 @rgiordan
paper.md attributions
https://github.com/openjournals/joss-reviews/issues/425

The review has the following checkbox, which is not fulfilled by the current paper.md: "Authors: Does the paper.md file include a list of authors with their affiliations?"  Somewhat relatedly, in my opinion, it might be made clearer in the paper that you're providing a ```shiny``` frontend to ```bnlearn``` rather than implementing the learning algorithms yourself.  Although you do very prominently say it's "powered" by ```bnlearn```, I feel it is best to be as clear as possible with author attribution!
