### 2017-10-20 #425 @rgiordan
https://github.com/paulgovan/BayesianNetwork/issues/7

The author and affiliation are missing from ```paper.md```, and I feel that ```bnlearn``` could be given clearer credit for the underlying learning algorithms.
