### 2017-10-20 paulgovan/BayesianNetwork#8 @rgiordan
paper missing references
"References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?"

The paper as-is (in md and pdf form) is entirely missing a references section (even though I see you have populated the bib file).
