### 2017-10-20 #425 @rgiordan
https://github.com/paulgovan/BayesianNetwork/issues/8

The paper is missing references.
