### 2017-10-20 #425 @rgiordan
I played around with the package, but there are many combinations of options so I didn't evaluate them exhaustively.  As such I'm not 100% sure what the standards are for "Functionality: Have the functional claims of the software been confirmed?"  I'll check it, but maybe @katyhuff can clarity the expectation are?

I would also feel more comfortable if there were unit tests.
