### 2017-10-20 #425 @rgiordan
@katyhuff @paulgovan

This is very nice work!  It's a beautiful and, I imagine, usable interface to a powerful set of learning algorithms.  I think it is well worth publishing.  However, the reviewing standards are quite clear -- it needs automated tests, and if they are in there now, I can't see them.  I'd recommend acceptance as long as good tests can be added.

Is there anything else you need me to fill out or do for the time being?
