### 2017-10-20 fb54626
Update paper.md
