### 2017-10-20 7b29b4d
Update paper.md
