### 2017-10-20 #425 @katyhuff
Thanks so much, @rgiordan, for your review. Great work!

The next step is for @paulgovan to look over the comments and issues you've created to make appropriate changes.

@rgiordan Regarding your question on the confirmation of the functional claims, we expect that the few claims of functionality in paper.md be checked by the reviewer to their satisfaction. This is where domain expertise comes in. For some packages and some reviewers, this means running an example, checking that a plot is as expected, or reviewing the test coverage. Indeed, as you point out @rgiordan, it is usually made much easier with unit tests and examples, so that's why we require unit tests.
