### 2017-10-20 65caa60
Update paper.md
