### 2017-10-20 paulgovan/BayesianNetwork#6 @paulgovan
Agreed. Unit tests are difficult in shiny, but recently the [shinytest](https://github.com/rstudio/shinytest) package was released, which looks like a possible solution. This may take a while, but I'll work to incorporate tests in the package.
