### 2017-10-20 paulgovan/BayesianNetwork#7 @paulgovan
Agreed as well. Added (fixed) affiliation and attempted to provide clear credit to the package.
