### 2017-10-20 paulgovan/BayesianNetwork#8 @paulgovan
Hmm I'm not terribly familiar with markdown, but I think I see the problem. Should be fixed shortly.
