### 2017-10-20 #425 @paulgovan
@rgiordan, @katyhuff

Please reference [#6](https://github.com/paulgovan/BayesianNetwork/issues/6), [#7](https://github.com/paulgovan/BayesianNetwork/issues/7), and [#8](https://github.com/paulgovan/BayesianNetwork/issues/8). Working on unit tests, but this could take several days.

Thank you for the thorough review!
