+++ inst/bn/ui.R
@@ -48,11 +48,14 @@ bnlearn::mb(dag, "A")
 
 ### Plot a d3 heatmap of the adjacency matrix
 ```{r}
-d3heatmap(
+heatmaply::heatmaply(
   bnlearn::amat(dag),
+  grid_gap = 1,
+  colors = blues9,
+  dendrogram = "both",
   symm = TRUE,
-  colors = "Blues",
-  opacity = 0.75
+  margins = c(100, 100, NA, 0),
+  hide_colorbar = TRUE
 )
 ````
 
@@ -99,7 +102,7 @@ shinydashboard::dashboardPage(
       # Parameters menu item
       shinydashboard::menuItem(
         "Parameters",
-        tabName = "paramaters",
+        tabName = "parameters",
         icon = shiny::icon("bar-chart")
       ),
 
@@ -221,24 +224,35 @@ shinydashboard::dashboardPage(
                                     width = NULL,
                                     shiny::helpText("Select a sample network or upload your Bayesian network data:"),
 
-                                    # Demo network input select
-                                    shiny::selectInput(
-                                      inputId = "net",
-                                      h5("Bayesian Network:"),
-                                      c("Sample Discrete Network" = 1,
-                                        "Sample Gaussian Network" = 2,
-                                        "Alarm Network" = 3,
-                                        "Insurance Network" = 4,
-                                        "Hailfinder Network" = 5,
-                                        "Upload your Bayesian network data" = 6
+                                    shinyWidgets::radioGroupButtons(inputId = "dataInput",
+                                                                    choices = c("Sample Network" = 1,
+                                                                                "Upload Data" = 2),
+                                                                    selected = 1,
+                                                                    justified = TRUE
+                                                                    ),
+
+                                    # Conditional panel for sample network selection
+                                    shiny::conditionalPanel(
+                                      condition = "input.dataInput == 1",
+
+                                      # Demo network input select
+                                      shiny::selectInput(
+                                        inputId = "net",
+                                        h5("Bayesian Network:"),
+                                        c("Sample Discrete Network" = 1,
+                                          "Sample Gaussian Network" = 2,
+                                          "Alarm Network" = 3,
+                                          "Insurance Network" = 4,
+                                          "Hailfinder Network" = 5
+                                        )
                                       )
                                     ),
 
                                     # Conditional panel for file input selection
                                     shiny::conditionalPanel(
-                                      condition = "input.net == 6",
-                                      shiny::p('Note: your data should be structured as a ',
-                                               shiny::a(href = 'http://en.wikipedia.org/wiki/Comma-separated_values', 'csv file')
+                                      condition = "input.dataInput == 2",
+
+                                      shiny::p('Note: your data should be structured as a csv file with header of variable names.'
                                       ),
 
                                       # File input
@@ -252,19 +266,6 @@ shinydashboard::dashboardPage(
                                                    '.csv',
                                                    '.tsv'
                                         )
-                                      ),
-
-                                      # Header T/F checkbox
-                                      shiny::checkboxInput('header', 'Header', TRUE),
-
-                                      # Separator input select
-                                      shiny::selectInput(
-                                        'sep',
-                                        shiny::strong('Separator:'),
-                                        c(Comma = ',',
-                                          Semicolon = ';',
-                                          Tab = '\t'
-                                        ), ','
                                       )
                                     )
                                   ),
@@ -347,21 +348,21 @@ shinydashboard::dashboardPage(
                               shiny::actionButton("structureIntro", "Show me how")
       ),
 
-      # Paramaters tab item
-      shinydashboard::tabItem(tabName = "paramaters",
+      # parameters tab item
+      shinydashboard::tabItem(tabName = "parameters",
                               shiny::fluidRow(
[…]
