+++ NAMESPACE
@@ -9,4 +9,5 @@ import(plotly)
 import(rintrojs)
 import(shiny)
 import(shinyAce)
+import(shinyWidgets)
 import(shinydashboard)
