### 2017-10-29 e65fc7f
Add shinytests
