+++ inst/bn/tests/mytest.R
@@ -1,0 +1,36 @@
+app <- ShinyDriver$new("../")
+app$snapshotInit("mytest")
+
+app$snapshot()
+app$setInputs(homeIntro = "click")
+app$setInputs(sidebarMenu = "structure")
+app$snapshot()
+app$setInputs(structureIntro = "click")
+app$setInputs(net = "3")
+app$setInputs(alg = "hc")
+app$setInputs(type = "aic")
+app$setInputs(sidebarMenu = "parameters")
+app$snapshot()
+app$setInputs(parametersIntro = "click")
+app$setInputs(met = "bayes")
+app$setInputs(param = "dotplot")
+app$setInputs(Node = "PCWP")
+app$setInputs(sidebarMenu = "inference")
+app$snapshot()
+app$setInputs(inferenceIntro = "click")
+app$setInputs(evidenceNode = "PCWP")
+app$setInputs(evidence = "HIGH")
+app$setInputs(event = "HIST")
+app$setInputs(sidebarMenu = "measures")
+app$snapshot()
+app$setInputs(measuresIntro = "click")
+app$setInputs(nodeMeasure = "nbr")
+app$setInputs(nodeNames = "PCWP")
+app$setInputs(dendrogram = "row")
+app$setInputs(sidebarMenu = "editor")
+app$snapshot()
+app$setInputs(editorIntro = "click")
+app$setInputs(bookmark = "click")
+app$setInputs(sidebarMenu = "structure")
+app$setInputs(dataInput = "2")
+app$uploadFile(file = "learning_test.csv")
