### 2017-10-29 ab3d1e6
Remove shinytest dependency
