### 2017-10-29 98bd675
Update tests
