### 2017-10-30 paulgovan/BayesianNetwork#6 @paulgovan
Ok I've added tests for the app's basic functionality via the `shinytest` package. `shinytest` is still under development, so the dev version must be installed from GitHub:

`
devtools::install_github("rstudio/shinytest")
`

To run the tests, navigate to the package directory and run:

`
shinytest::testApp("inst/bn")
`

There may be some slight differences between some of the image snapshots due to the randomness of the d3 graphics, but the tests should pass otherwise.
