### 2017-10-30 #425 @paulgovan
@rgiordan, @katyhuff

Please see my changes referenced in the issues above. Let me know if you have any questions.
