### 2017-11-20 #425 @katyhuff
Ah!! Sorry, I was so confused -- the issues are still open! If they have been fixed, it's appropriate to close them (only you can.. we don't own this repository). I'm looking now to ensure that I can run the tests.
