### 2017-11-20 paulgovan/BayesianNetwork#7 @katyhuff
Just a note, in order to communicate that something is fixed, it's common to comment as you have above. However, that comment often includes a link to either a PR, or the merged commit that fixes the issue. In this case, I think it was probably this commit (https://github.com/paulgovan/BayesianNetwork/commit/fb54626f85ffc6a2f7378f951561d73d7400aea8#diff-60426898a850f648abeda63b199792cc) that solved the issue.
