### 2017-11-20 paulgovan/BayesianNetwork#8 @katyhuff
Hi @paulgovan . This still isn't working. Have you tried actually building the paper? (use pandoc). Your paper doesn't seem to actually cite any of the references in your .bib file. Thus, none appear in the list of references. You need to cite references thus `Previous authors [@scutari2009learning] ...`.

Relatedly, it is only possible to do this citation if your bib file has keys for each of the entries (e.g. scutari2009learning) many are missing a key, which should be after the curly brace and before the comma... .
