### 2017-11-20 #425 @katyhuff
- [ ] The issue with the paper references hasn't been fixed. Much like an ordinary paper, you must actually cite things in the text in order for them to appear in the references.
- [ ] I strongly recommend a caption for the figure in your paper.
- [ ] It's not clear that your unit tests are robustly testing the method itself.
- [ ] We require either automated tests (e.g. with travis) or instructions on how to run the tests ( in the readme, for future users, not just for us). Your travis instance currently doesn't run your new shiny tests. I recommend adding those commands to your travis.yml file. Once that's done, travis will automatically run your tests. If you'd rather not, then please at least include instructions in the readme on how a user can run your tests.
