### 2017-11-20 paulgovan/BayesianNetwork#8 @paulgovan
Ah ok I see. Ok I will actually cite each reference in the paper and add keys to all of them in the .bib file.
