### 2017-11-20 paulgovan/BayesianNetwork#7 @paulgovan
Yes, please reference commits [https://github.com/paulgovan/BayesianNetwork/commit/fb54626f85ffc6a2f7378f951561d73d7400aea8](https://github.com/paulgovan/BayesianNetwork/commit/fb54626f85ffc6a2f7378f951561d73d7400aea8) and [https://github.com/paulgovan/BayesianNetwork/commit/7b29b4df7aa5d947822b5dee43a9a77dda34545b](https://github.com/paulgovan/BayesianNetwork/commit/7b29b4df7aa5d947822b5dee43a9a77dda34545b) on this fix.
