### 2017-11-21 6b140c0
Update paper references
