+++ inst/paper/paper.bib
@@ -1,4 +1,4 @@
-  @Manual{,
+  @Manual{BayesianNetwork,
     title = {BayesianNetwork: Bayesian Network Modeling and Analysis},
     author = {Paul Govan},
     year = {2017},
@@ -21,7 +21,7 @@
     year={2009}
   }
 
-  @Manual{,
+  @Manual{shiny,
     title = {shiny: Web Application Framework for R},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
     year = {2017},
@@ -29,7 +29,7 @@
     url = {https://CRAN.R-project.org/package=shiny},
   }
 
-  @Manual{,
+  @Manual{shinydashboard,
     title = {shinydashboard: Create Dashboards with 'Shiny'},
     author = {Winston Chang and Barbara {Borges Ribeiro}},
     year = {2017},
@@ -37,7 +37,7 @@
     url = {https://CRAN.R-project.org/package=shinydashboard},
   }
 
-  @Manual{,
+  @Manual{heatmaply,
     title = {heatmaply: Interactive Cluster Heat Maps Using 'plotly'},
     author = {Tal Galili},
     year = {2017},
@@ -45,7 +45,7 @@
     url = {https://CRAN.R-project.org/package=heatmaply},
   }
 
-  @Manual{,
+  @Manual{networkD3,
     title = {networkD3: D3 JavaScript Network Graphs from R},
     author = {J.J. Allaire and Christopher Gandrud and Kenton Russell and CJ Yetman},
     year = {2017},
@@ -53,7 +53,7 @@
     url = {https://CRAN.R-project.org/package=networkD3},
   }
 
-  @Article{,
+  @Article{rintrojs,
     title = {{rintrojs}: A Wrapper for the Intro.js Library},
     author = {Carl Ganz},
     journal = {Journal of Open Source Software},
@@ -65,7 +65,7 @@
     url = {http://dx.doi.org/10.21105/joss.00063},
   }
 
-  @Manual{,
+  @Manual{shinyAce,
     title = {shinyAce: Ace Editor Bindings for Shiny},
     author = {{Trestle Technology, LLC}},
     year = {2016},
