### 2017-11-21 paulgovan/BayesianNetwork#8 @paulgovan
I found that there are at least 2 ways to ensure a reference is included in the References section: an in text citation or use of the `nocite` metadata field. The references appear to be rendering now. Please see commit [https://github.com/paulgovan/BayesianNetwork/commit/6b140c00ea059ed2fa641627be809d855295e217](https://github.com/paulgovan/BayesianNetwork/commit/6b140c00ea059ed2fa641627be809d855295e217).
