### 2017-11-26 7478df5
Automate tests
