+++ inst/bn/tests/mytest.R
@@ -30,10 +30,6 @@ Sys.sleep(1)
 app$setInputs(nodeMeasure = "nbr")
 app$setInputs(nodeNames = "PCWP")
 app$setInputs(dendrogram = "row")
-app$setInputs(sidebarMenu = "editor")
-Sys.sleep(5)
-app$snapshot()
-Sys.sleep(1)
 app$setInputs(sidebarMenu = "structure")
 app$setInputs(dataInput = "2")
 app$uploadFile(file = "learning_test.csv")
