+++ run_tests.R
@@ -1,0 +1,3 @@
+testthat::test_that("Application works", {
+  shinytest::expect_pass(shinytest::testApp("inst/bn", "mytest", compareImages = FALSE))
+})
