### 2017-11-26 c085eb6
Automate tests
