+++ DESCRIPTION
@@ -24,7 +24,7 @@ Imports:
     shinyWidgets,
     testthat
 Remotes:
-    rstudio/shinytest
+    rstudio/shinytest@dev
 RoxygenNote: 6.0.1
 Suggests: knitr,
     rmarkdown
