+++ .travis.yml
@@ -1,4 +1,4 @@
-dist: trusty   # Use Ubuntu 14.04 image (instead of 12.04)
+# dist: trusty   # Use Ubuntu 14.04 image (instead of 12.04)
 language: r
 sudo: false
 r: 3.4.1
