### 2017-11-26 74163c5
Automate tests
