### 2017-11-26 2ca279f
Update DESCRIPTION
