+++ DESCRIPTION
@@ -13,7 +13,6 @@ License: Apache License | file LICENSE
 LazyData: TRUE
 Imports:
     bnlearn,
-    Cairo,
     heatmaply,
     lattice,
     networkD3,
