### 2017-12-03 c55e35a
Update tests
