+++ run_tests.R
@@ -1,4 +1,5 @@
-install.packages("Cairo", INSTALL_opts = "--no-test-load")
+library(testthat)
+library(shinytest)
 
 testthat::test_that("Application works", {
   shinytest::expect_pass(shinytest::testApp("inst/bn", "mytest", compareImages = FALSE))
