+++ .travis.yml
@@ -1,7 +1,7 @@
 # dist: trusty   # Use Ubuntu 14.04 image (instead of 12.04)
 language: r
 sudo: false
-# r: 3.4.1
+r: 3.4.1
 
 cache:
   # Main R library
