+++ run_tests.R
@@ -2,5 +2,5 @@ library(testthat)
 library(shinytest)
 
 testthat::test_that("Application works", {
-  shinytest::expect_pass(shinytest::testApp("inst/bn", "mytest", compareImages = FALSE))
+  shinytest::expect_pass(shinytest::testApp("inst/bn", "mytest", quiet = TRUE, compareImages = FALSE))
 })
