+++ inst/bn/tests/mytest.R
@@ -3,37 +3,33 @@ app$snapshotInit("mytest")
 
 Sys.sleep(1)
 app$snapshot()
-Sys.sleep(2)
+Sys.sleep(3)
 app$setInputs(sidebarMenu = "structure")
 Sys.sleep(5)
 app$snapshot()
-Sys.sleep(1)
-app$setInputs(net = "3")
 app$setInputs(alg = "hc")
 app$setInputs(type = "aic")
 app$setInputs(sidebarMenu = "parameters")
-app$snapshot()
 Sys.sleep(1)
+app$snapshot()
 app$setInputs(met = "bayes")
 app$setInputs(param = "dotplot")
-app$setInputs(Node = "PCWP")
+app$setInputs(Node = "B")
 app$setInputs(sidebarMenu = "inference")
-app$snapshot()
 Sys.sleep(1)
-app$setInputs(evidenceNode = "PCWP")
-app$setInputs(evidence = "HIGH")
-app$setInputs(event = "HIST")
+app$snapshot()
+app$setInputs(evidenceNode = "B")
+app$setInputs(evidence = "a")
+app$setInputs(event = "C")
 app$setInputs(sidebarMenu = "measures")
-Sys.sleep(5)
-app$snapshot()
 Sys.sleep(1)
+app$snapshot()
 app$setInputs(nodeMeasure = "nbr")
-app$setInputs(nodeNames = "PCWP")
+app$setInputs(nodeNames = "B")
 app$setInputs(dendrogram = "row")
+app$setInputs(sidebarMenu = "editor")
+app$setInputs(bookmark = "click")
 app$setInputs(sidebarMenu = "structure")
+app$setInputs(net = "3")
 app$setInputs(dataInput = "2")
 app$uploadFile(file = "learning_test.csv")
-Sys.sleep(5)
-app$snapshot()
-Sys.sleep(1)
-app$setInputs(bookmark = "click")
