### 2017-12-03 d53283c
Update tests
