+++ inst/bn/tests/mytest.R
@@ -1,12 +1,12 @@
-app <- ShinyDriver$new("../", seed = 1)
+app <- ShinyDriver$new("../", seed = 123)
 app$snapshotInit("mytest")
 
-Sys.sleep(1)
 app$snapshot()
-Sys.sleep(3)
+Sys.sleep(2)
 app$setInputs(sidebarMenu = "structure")
-Sys.sleep(5)
+Sys.sleep(4)
 app$snapshot()
+app$setInputs(net = "3")
 app$setInputs(alg = "hc")
 app$setInputs(type = "aic")
 app$setInputs(sidebarMenu = "parameters")
@@ -14,22 +14,20 @@ Sys.sleep(1)
 app$snapshot()
 app$setInputs(met = "bayes")
 app$setInputs(param = "dotplot")
-app$setInputs(Node = "B")
+app$setInputs(Node = "PCWP")
 app$setInputs(sidebarMenu = "inference")
 Sys.sleep(1)
 app$snapshot()
-app$setInputs(evidenceNode = "B")
-app$setInputs(evidence = "a")
-app$setInputs(event = "C")
+app$setInputs(evidenceNode = "PCWP")
+app$setInputs(evidence = "HIGH")
+app$setInputs(event = "HIST")
 app$setInputs(sidebarMenu = "measures")
-Sys.sleep(1)
-app$snapshot()
 app$setInputs(nodeMeasure = "nbr")
-app$setInputs(nodeNames = "B")
+app$setInputs(nodeMeasure = "parents")
+app$setInputs(nodeNames = "PCWP")
 app$setInputs(dendrogram = "row")
 app$setInputs(sidebarMenu = "editor")
 app$setInputs(bookmark = "click")
 app$setInputs(sidebarMenu = "structure")
-app$setInputs(net = "3")
 app$setInputs(dataInput = "2")
 app$uploadFile(file = "learning_test.csv")
