### 2017-12-03 paulgovan/BayesianNetwork#6 @paulgovan
Ok I've integrated testing with travis as of commit [https://github.com/paulgovan/BayesianNetwork/commit/d53283cfad5b7001e161f6c762f82bca4a19edac](https://github.com/paulgovan/BayesianNetwork/commit/d53283cfad5b7001e161f6c762f82bca4a19edac). Note that this testing framework is still under development, and I will be looking to add more details in the future.
