### 2017-12-03 #425 @paulgovan
@rgiordan, @katyhuff

I believe that issues [#6](https://github.com/paulgovan/BayesianNetwork/issues/6), [#7](https://github.com/paulgovan/BayesianNetwork/issues/7), and [#8](https://github.com/paulgovan/BayesianNetwork/issues/8) have been fixed, and they are now closed.
