### 2018-01-04 #425 @katyhuff
@paulgovan This looks good to me! I believe it can be accepted! At this point could you make an updated archive of the reviewed version of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
